+++ src/Copper/Copper.php
@@ -34,8 +34,6 @@ class Copper
 
     /**
      * Instance of self.
-     *
-     * @var Copper/Copper|null
      */
     public static ?Copper $instance = null;
 
@@ -64,7 +62,7 @@ class Copper
      */
     public static function create(float|string|null $value = null, ?int $style = null, ?string $locale = null): Copper
     {
-        if (! self::$instance instanceof \Copper\Copper) {
+        if (! self::$instance instanceof self) {
             self::$instance = new self;
         }
 
@@ -87,9 +85,9 @@ class Copper
      * Format the number using DECIMAL.
      *
      * @param  int|null  $precision  Precision to use for formatter.
-     * @return string Formatted number.
+     * @return string|bool Formatted number.
      */
-    public static function decimal(?int $precision = null): string
+    public static function decimal(?int $precision = null): string|bool
     {
         if (false === is_null($precision)) {
             self::$formatter->setAttribute(NumberFormatter::FRACTION_DIGITS, $precision);
@@ -105,9 +103,9 @@ class Copper
      * Format the number using CURRENCY.
      *
      * @param  string  $iso  The 3-letter ISO 4217 currency code indicating the currency to use.
-     * @return string Formatted number.
+     * @return string|bool Formatted number.
      */
-    public static function currency(string $iso): string
+    public static function currency(string $iso): string|bool
     {
         if (NumberFormatter::CURRENCY !== self::$style) {
             self::setStyle(NumberFormatter::CURRENCY);
@@ -119,9 +117,9 @@ class Copper
     /**
      * Format the number using SPELLOUT.
      *
-     * @return string Formatted number.
+     * @return string|bool Formatted number.
      */
-    public static function spellOut(): string
+    public static function spellOut(): string|bool
     {
         if (NumberFormatter::SPELLOUT !== self::$style) {
             self::setStyle(NumberFormatter::SPELLOUT);
@@ -133,9 +131,9 @@ class Copper
     /**
      * Format the number using PERCENT.
      *
-     * @return string Formatted number.
+     * @return string|bool Formatted number.
      */
-    public static function percentage(): string
+    public static function percentage(): string|bool
     {
         if (NumberFormatter::PERCENT !== self::$style) {
             self::setStyle(NumberFormatter::PERCENT);
@@ -148,9 +146,9 @@ class Copper
      * Format the number using CURRENCY_ACCOUNTING.
      *
      * @param  string  $iso  The 3-letter ISO 4217 currency code indicating the currency to use.
-     * @return string Formatted number.
+     * @return string|bool Formatted number.
      */
-    public static function accounting(string $iso): string
+    public static function accounting(string $iso): string|bool
     {
         if (NumberFormatter::CURRENCY_ACCOUNTING !== self::$style) {
             self::setStyle(NumberFormatter::CURRENCY_ACCOUNTING);
@@ -162,9 +160,9 @@ class Copper
     /**
      * Format the number using SCIENTIFIC.
      *
-     * @return string Formatted number.
+     * @return string|bool Formatted number.
      */
-    public static function scientific(): string
+    public static function scientific(): string|bool
     {
         if (NumberFormatter::SCIENTIFIC !== self::$style) {
             self::setStyle(NumberFormatter::SCIENTIFIC);
@@ -220,9 +218,9 @@ class Copper
      * Set the Locale.
      *
      * @param  string  $locale  Locale in which the number would be formatted.
-     * @return Copper Copper instance.
+     * @return ?Copper Copper instance.
      */
-    public static function setLocale(string $locale): Copper
+    public static function setLocale(string $locale): ?Copper
     {
         self::$locale = $locale;
         self::create();
@@ -244,9 +242,9 @@ class Copper
      * Set the Style.
      *
      * @param  int  $style  Style of the formatting.
-     * @return Copper Copper instance.
+     * @return ?Copper Copper instance.
      */
-    public static function setStyle(int $style): Copper
+    public static function setStyle(int $style): ?Copper
     {
         self::$style = $style;
         self::create();
